PluginHandler: shrink interface.
Move some specific filter functions to the pluginmanager, the
only place they are used.
