+++ src/PluginHandler.cpp
@@ -689,33 +689,6 @@ void PluginHandler::cleanup(const std::string& filelist,
 }
 
 
-std::vector<PluginMetadata> PluginHandler::getUpdates(const std::string& name)
-{
-    auto av = getAvailable();
-    // Skip all available which doesnt refer to correct name
-    av.erase(
-        std::remove_if(
-            av.begin(), av.end(),
-            [&](const PluginMetadata& pm) {return pm.name != name; }),
-        av.end());
-    //  Get metadata for installed version (if existing).
-    auto in = getInstalled();
-    auto current = std::find_if(
-            in.begin(), in.end(),
-            [&] (const PluginMetadata& pm) { return pm.name == name; });
-    if (current == in.end()) {
-        return av;
-    }
-    // Drop item matching installed version.
-    av.erase(
-        std::remove_if(av.begin(), av.end(),
-                       [&](const PluginMetadata& pm) {
-                            return pm.version == current->version; }),
-        av.end());
-    return av;
-}
-
-
 const std::vector<PluginMetadata> PluginHandler::getAvailable()
 {
     using namespace std;
@@ -863,29 +836,4 @@ bool PluginHandler::uninstall(const std::string plugin_name)
     remove(PluginHandler::versionPath(plugin_name).c_str());  // are OK.
 
     return true;
-}
-
-std::vector<PluginMetadata> PluginHandler::getAvailableUniquePlugins()
-{
-    /** Compare two PluginMetadata objects, a named c++ requirement. */
-    struct metadata_compare{
-        bool operator() (const PluginMetadata& lhs,
-                         const PluginMetadata& rhs) const
-        {
-            return lhs.key() < rhs.key();
-        }
-    };
-
-    std::vector<PluginMetadata> returnArray;
-    
-    std::set<PluginMetadata, metadata_compare> unique_plugins;
-    for (auto plugin: getAvailable()) {
-        unique_plugins.insert(plugin);
-    }
-    for (auto plugin: unique_plugins) {
-        if (PluginHandler::isCompatible(plugin)) {
-            returnArray.push_back(plugin);
-        }
-    }
-    return returnArray;
 }
